change dir format
