+++ scripts/download/tripleca/01_download_3ca_run_download_3ca.sh
@@ -14,7 +14,7 @@ mkdir -p "$DATA_PATH_3CA"
 mkdir -p "$LOG_PATH"
 
 # 运行 Python 下载脚本
-python -m scripts.download.tripleca.download_3ca \
+python -m src.deepsc.data.download.tripleca.download_3ca \
     --output_path "$DATA_PATH_3CA" \
     --log_path "$LOG_PATH" \
     --num_files "$NUM_FILES" \
