+++ scripts/download/tripleca/02_download_3ca_run_filter_dataset.sh
@@ -6,5 +6,5 @@ set -a
 source .env
 set +a
 
-python -m scripts.download.tripleca.merge_and_filter_dataset \
+python -m src.deepsc.data.download.tripleca.merge_and_filter_dataset \
     --dataset_root_path "$DATA_PATH_3CA"
