+++ src/deepsc/data/download/cellxgene/download_partition.py
