+++ src/deepsc/data/download/tripleca/config.py
