+++ src/deepsc/data/preprocessing/preprocess_datasets_cellxgene.py
@@ -5,9 +5,8 @@ from functools import partial
 from tqdm import tqdm
 
 from multiprocessing import Pool, cpu_count
-from scripts.preprocessing.preprocess_datasets import process_h5ad_to_sparse_tensor
-from scripts.utils.utils import setup_logging
-
+from deepsc.data.preprocessing.preprocess_datasets import process_h5ad_to_sparse_tensor
+from deepsc.utils import setup_logging
 
 def find_all_h5ad_files(root_dir: str):
     h5ad_files = []
