+++ src/deepsc/models/deepsc/deepsc.py
@@ -1,0 +1,2 @@
+def printDSC():
+    print("#######################")
