+++ src/deepsc/models/scbert/__init__.py
@@ -1,0 +1,1 @@
+from deepsc.models.scbert.scbert import *
