+++ src/deepsc/test.py
@@ -1,0 +1,3 @@
+from src.deepsc.models.deepsc.deepsc import printDSC
+
+printDSC()
