+++ src/deepsc/utils/__init__.py
@@ -1,0 +1,1 @@
+from .utils import *
