--- src/scBERT/.gitignore
@@ -1,24 +1,0 @@
-
- *.log
- *.txt
- # 忽略 node_modules 目录
- node_modules/
- test.py
- test.sh
- performer_pytorch/
- # 忽略 Python 的虚拟环境
- venv/
- core.*
- # 忽略 Mac 系统的特殊文件
- .DS_Store
- 
- # 忽略所有以 ~ 结尾的临时文件
- *~
- __pycache__/
- ckpts/
- # 忽略特定配置文件
- config.json
- 
- # 可以通过 ! 将某些文件从忽略中排除
- !important.log
- preprocess_cellxgene.py
