--- src/scBERT/performer_pytorch/__init__.py
@@ -1,6 +1,0 @@
-from performer_pytorch.performer_pytorch import (
-    FastAttention,
-    Performer,
-    PerformerLM,
-    SelfAttention,
-)
