--- src/scBERT/pretrain.py
@@ -1,370 +1,0 @@
-
-# -*- coding: utf-8 -*-
-import os
-import argparse
-import random
-import logging
-import os.path as osp
-import math
-import random
-from functools import reduce
-from sklearn.model_selection import train_test_split
-import torch
-from torch import nn
-from torch.optim import Adam
-from torch.utils.data import DataLoader, Dataset
-from torch.utils.data.distributed import DistributedSampler
-import torch.distributed as dist
-from performer_pytorch import PerformerLM
-import anndata as ad
-from utils import *
-from lightning.fabric import Fabric
-
-def setup_logging(type, log_path):
-    os.makedirs(log_path, exist_ok=True)
-
-    timestamp = 'today'
-    log_file = osp.join(log_path, f"pretrain_{type}_{timestamp}.log")
-
-    logging.basicConfig(
-        filename=log_file,
-        level=logging.INFO,
-        format="%(asctime)s - %(levelname)s - %(message)s",
-        filemode="w",
-    )
-    console = logging.StreamHandler()
-    console.setLevel(logging.INFO)
-    formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
-    console.setFormatter(formatter)
-    logging.getLogger("").addHandler(console)
-
-    logging.info(f"日志文件: {log_file}")
-
-    return log_file
-
-
-
-
-parser = argparse.ArgumentParser()
-parser.add_argument("--local_rank", type=int, default=-1, help='Local process rank.')
-parser.add_argument("--local-rank", type=int, default=-1, help='Local process rank.')
-parser.add_argument("--bin_num", type=int, default=5, help='Number of bins.')
-parser.add_argument("--gene_num", type=int, default=60664, help='Number of genes.')
-parser.add_argument("--epoch", type=int, default=1, help='Number of epochs.')
-parser.add_argument("--seed", type=int, default=2021, help='Random seed.')
-parser.add_argument("--batch_size", type=int, default=3, help='Number of batch size.')
-parser.add_argument("--learning_rate", type=float, default=1e-4, help='Learning rate.')
-parser.add_argument("--grad_acc", type=int, default=60, help='Number of gradient accumulation.')
-parser.add_argument("--valid_every", type=int, default=1, help='Number of training epochs between twice validation.')
-parser.add_argument("--mask_prob", type=float, default=0.15, help='Probability of masking.')
-parser.add_argument("--replace_prob", type=float, default=0.9, help='Probability of replacing with [MASK] token for masking.')
-parser.add_argument("--pos_embed", type=bool, default=True, help='Using Gene2vec encoding or not.')
-parser.add_argument("--data_path", type=str, default='/hpcwork/p0021245/smallset', help='Path of data for pretraining.')
-parser.add_argument("--ckpt_dir", type=str, default='./ckpts/', help='Directory of checkpoint to save.')
-parser.add_argument("--model_name", type=str, default='cellxgene_model_no_divide', help='Pretrained model name.')
-args = parser.parse_args()
-local_rank = args.local_rank
-print(f"local_rank: {local_rank}")
-rank = int(os.environ["RANK"])
-is_master = rank == 0
-master_addr = os.getenv("MASTER_ADDR", "127.0.0.1")
-master_port = os.getenv("MASTER_PORT", "12345")
-rank = int(os.getenv("SLURM_PROCID", "0"))
-world_size = int(os.getenv("SLURM_NTASKS", "1"))
-local_rank = int(os.getenv("SLURM_LOCALID"))
-setup_logging(rank, './')
-SEED = args.seed
-EPOCHS = args.epoch
-BATCH_SIZE = args.batch_size
-GRADIENT_ACCUMULATION = args.grad_acc
-LEARNING_RATE = args.learning_rate
-SEQ_LEN = args.gene_num + 1
-VALIDATE_EVERY = args.valid_every
-CLASS = args.bin_num + 2
-MASK_PROB = args.mask_prob
-REPLACE_PROB = args.replace_prob
-RANDOM_TOKEN_PROB = 0.
-MASK_TOKEN_ID = CLASS - 1
-PAD_TOKEN_ID = CLASS - 1
-MASK_IGNORE_TOKEN_IDS = [0]
-POS_EMBED_USING = args.pos_embed
-model_name = args.model_name
-ckpt_dir = args.ckpt_dir
-
-
-dist.init_process_group(
-    backend="nccl",
-    init_method=f"tcp://{master_addr}:{master_port}",
-    rank=rank,
-    world_size=world_size
-)
-torch.cuda.set_device(local_rank)
-device = torch.device("cuda", local_rank)
-world_size = torch.distributed.get_world_size()
-seed_all(SEED + torch.distributed.get_rank())
-
-# get the random prob matrix and True means smaller than prob threshold
-def prob_mask_like(t, prob):
-    return torch.zeros_like(t).float().uniform_(0, 1) < prob
-
-# get the mask matrix which cannot be masked
-def mask_with_tokens(t, token_ids):
-    init_no_mask = torch.full_like(t, False, dtype=torch.bool)
-    mask = reduce(lambda acc, el: acc | (t == el), token_ids, init_no_mask)
-    return mask
-
-def get_mask_subset_with_prob(mask, prob):
-    batch, seq_len, device = *mask.shape, mask.device
-    max_masked = math.ceil(prob * seq_len)      # num of mask of a single sequence in average
-    num_tokens = mask.sum(dim=-1, keepdim=True)     # num of pure tokens of each sequence except special tokens
-    mask_excess = torch.cat((torch.zeros(0), torch.arange(mask.size(-1)).repeat(mask.size(0)))).reshape(mask.size(0),mask.size(-1)).to(device)
-    mask_excess = (mask_excess >= (num_tokens * prob).ceil())        # only 15% of pure tokens can be masked
-    mask_excess = mask_excess[:, :max_masked]       # get difference between 15% of pure tokens and 15% of all tokens
-    rand = torch.rand((batch, seq_len), device=device).masked_fill(~mask, -1e9)     # rand (0-1) as prob, special token use -1e9
-    _, sampled_indices = rand.topk(max_masked, dim=-1)      # get index of topk prob to mask
-    sampled_indices = (sampled_indices + 1).masked_fill_(mask_excess, 0)        # delete difference of mask not pure
-    new_mask = torch.zeros((batch, seq_len + 1), device=device)     # get (batch, seq_len) shape zero matrix
-    new_mask.scatter_(-1, sampled_indices, 1)       # set masks in zero matrix as 1
-    return new_mask[:, 1:].bool()       # the final mask, True is mask
-
-def data_mask(data,
-    mask_prob = MASK_PROB,
-    replace_prob = REPLACE_PROB,
-    num_tokens = None,
-    random_token_prob = RANDOM_TOKEN_PROB,
-    mask_token_id = MASK_TOKEN_ID,
-    pad_token_id = PAD_TOKEN_ID,
-    mask_ignore_token_ids = MASK_IGNORE_TOKEN_IDS
-):
-    mask_ignore_token_ids = set([*mask_ignore_token_ids, pad_token_id])
-    # do not mask [pad] tokens, or any other tokens in the tokens designated to be excluded ([cls], [sep])
-    # also do not include these special tokens in the tokens chosen at random
-    no_mask = mask_with_tokens(data, mask_ignore_token_ids)   # ignore_token as True, will not be masked later
-    mask = get_mask_subset_with_prob(~no_mask, mask_prob)      # get the True/False mask matrix
-    # get mask indices
-    ## mask_indices = torch.nonzero(mask, as_tuple=True)   # get the index of mask(nonzero value of mask matrix)
-    # mask input with mask tokens with probability of `replace_prob` (keep tokens the same with probability 1 - replace_prob)
-    masked_input = data.clone().detach()
-    # if random token probability > 0 for mlm
-    if random_token_prob > 0:
-        assert num_tokens is not None, 'num_tokens keyword must be supplied when instantiating MLM if using random token replacement'
-        random_token_prob = prob_mask_like(data, random_token_prob)       # get the mask matrix of random token replace
-        random_tokens = torch.randint(0, num_tokens, data.shape, device=data.device)     # generate random token matrix with the same shape as input
-        random_no_mask = mask_with_tokens(random_tokens, mask_ignore_token_ids)        # not masked matrix for the random token matrix
-        random_token_prob &= ~random_no_mask        # get the pure mask matrix of random token replace
-        random_indices = torch.nonzero(random_token_prob, as_tuple=True)        # index of random token replace
-        masked_input[random_indices] = random_tokens[random_indices]        # replace some tokens by random token
-    # [mask] input
-    replace_prob = prob_mask_like(data, replace_prob)     # get the mask matrix of token being masked
-    masked_input = masked_input.masked_fill(mask * replace_prob, mask_token_id)        # get the data has been masked by mask_token
-    # mask out any tokens to padding tokens that were not originally going to be masked
-    labels = data.masked_fill(~mask, pad_token_id)        # the label of masked tokens
-    return masked_input, labels
-
-""" class SCDataset(Dataset):
-    def __init__(self, data):
-        super().__init__()
-        self.data = data
-
-    def __getitem__(self, index):
-        rand_start = random.randint(0, self.data.shape[0]-1)
-        full_seq = self.data[rand_start].toarray()[0]
-        full_seq[full_seq > (CLASS - 2)] = CLASS - 2
-        full_seq = torch.from_numpy(full_seq).long()
-        full_seq = torch.cat((full_seq, torch.tensor([0]))).to(device)
-        return full_seq
-
-    def __len__(self):
-        return self.data.shape[0] """
-class SCDataset(Dataset):
-    def __init__(self, coo_tensor):
-        if not coo_tensor.is_coalesced():
-            raise ValueError("输入的稀疏张量不是 coalesced 格式，请先调用 coalesce() 进行合并")
-        self.num_samples = coo_tensor.shape[0]
-    
-    def __getitem__(self, idx):
-        rand_idx = random.randint(0, self.num_samples - 1)
-        row = self.coo_tensor[rand_idx].to_dense()
-        row[row > (CLASS - 2)] = CLASS - 2
-        row = row.long()
-        row = torch.cat((row, torch.tensor([0]))).to(row.device)
-        return row
-
-    def __len__(self):
-        return self.num_samples
-
-""" def merge_h5ad_X(path_list):
-
-    anndata_list = []
-    for path in path_list:
-        data = ad.read_h5ad(path)
-        logging.info(f'Reading file: {path}')
-        anndata_list.append(data)
-        logging.info(f'File shape: {data.shape[0]}')
-    combined_anndata = ad.concat(anndata_list, axis=0)
-    logging.info(f'Combined shape: {combined_anndata.shape[0]}')
-
-    return combined_anndata.X
-
-def get_path_h5ad_in_dir(path):
-    path_list = []
-    for filename in os.listdir(path):
-            logging.info(f'filename: {filename}')
-            if filename.endswith(".h5ad"):
-                file_path = osp.join(path, filename)
-                path_list.append(file_path)
-    if len(path_list) == 0:
-        raise ValueError("no datasets available in the directory!")
-    return path_list """
-
-""" path_list = get_path_h5ad_in_dir(args.data_path)
-data = merge_h5ad_X(path_list) 
-
-data_train, data_val = train_test_split(data, test_size=0.05, random_state=SEED)
- """
-"用于提取目标行的稀疏张量"
-def extract_rows_from_sparse_tensor(tensor, row_ids):
-    tensor = tensor.coalesce()
-    idx = tensor.indices()
-    val = tensor.values()
-    
-    # 找出属于目标 row 的位置
-    mask = torch.isin(idx[0], torch.tensor(row_ids, device=idx.device))
-    new_indices = idx[:, mask]
-    new_values = val[mask]
-    
-    # 行号需要重新映射：比如原来 row 100 在 row_ids 中是第 5 行
-    row_id_map = {orig: i for i, orig in enumerate(row_ids)}
-    remapped_rows = torch.tensor([row_id_map[int(r)] for r in new_indices[0].tolist()], device=idx.device)
-    new_indices[0] = remapped_rows
-
-    return torch.sparse_coo_tensor(new_indices, new_values, size=(len(row_ids), tensor.shape[1]))
-
-tensor_path = "/hpcwork/p0021245/Data/test_pth_folder/1d84333c-0327-4ad6-be02-94fee81154ff_sparse_preprocessed.pth"
-coo_tensor = torch.load(tensor_path)
-
-num_rows = coo_tensor.shape[0]
-row_indices = np.arange(num_rows)
-train_idx, val_idx = train_test_split(row_indices, test_size=0.05, random_state=SEED)
-
-data_train = extract_rows_from_sparse_tensor(coo_tensor, train_idx)
-data_val = extract_rows_from_sparse_tensor(coo_tensor, val_idx)
-
-train_dataset = SCDataset(data_train)
-val_dataset = SCDataset(data_val)
-
-train_sampler = DistributedSampler(train_dataset)
-val_sampler = SequentialDistributedSampler(val_dataset, batch_size=BATCH_SIZE, world_size=world_size)
-
-train_loader = DataLoader(train_dataset, batch_size=BATCH_SIZE, sampler=train_sampler)
-val_loader = DataLoader(val_dataset, batch_size=BATCH_SIZE, sampler=val_sampler)
-
-fabric = Fabric(accelerator="cuda", devices=4, num_nodes=2, strategy="ddp")
-fabric.launch()
-
-train_loader, val_loader = fabric.setup_dataloaders(train_loader, val_loader)
-
-model = PerformerLM(
-    num_tokens = CLASS,
-    dim = 200,
-    depth = 6,
-    max_seq_len = SEQ_LEN,
-    heads = 10,
-    local_attn_heads = 0,
-    g2v_position_emb = POS_EMBED_USING
-)
-
-current_device = torch.cuda.current_device()
-
-# optimizer
-optimizer = Adam(model.parameters(), lr=LEARNING_RATE)
-model, optimizer = fabric.setup(model, optimizer)
-# learning rate scheduler
-scheduler = CosineAnnealingWarmupRestarts(
-    optimizer,
-    first_cycle_steps=15,
-    cycle_mult=2,
-    max_lr=LEARNING_RATE,
-    min_lr=1e-6,
-    warmup_steps=5,
-    gamma=0.9
-)
-loss_fn = nn.CrossEntropyLoss(ignore_index = PAD_TOKEN_ID, reduction='mean').to(local_rank)
-softmax = nn.Softmax(dim=-1)
-for i in range(1, EPOCHS+1):
-    train_loader.sampler.set_epoch(i)
-    model.train()
-    running_loss = 0.0
-    cum_acc = 0.0
-    for index, data in enumerate(train_loader):
-        index += 1
-        logging.info(f"Accumulated gradient on rank: {rank}, local_rank: {local_rank}")
-        data, labels = data_mask(data)
-        if index % GRADIENT_ACCUMULATION != 0:
-            with fabric.no_backward_sync(model, enabled=False):
-                logits = model(data)
-                loss = (loss_fn(logits.transpose(1, 2), labels) / GRADIENT_ACCUMULATION)
-                fabric.backward(loss)  
-
-        if index % GRADIENT_ACCUMULATION == 0:
-            logging.info(f"Accumulated gradient on rank: {rank}, local_rank: {local_rank}")
-            logits = model(data)
-            loss = loss_fn(logits.transpose(1, 2), labels) / GRADIENT_ACCUMULATION
-            fabric.backward(loss)
-            torch.nn.utils.clip_grad_norm_(model.parameters(), int(1e2))
-            optimizer.step()
-            optimizer.zero_grad()
-        running_loss += loss.item()
-        final = softmax(logits)[..., 1:-1]
-        final = final.argmax(dim=-1) + 1
-        pred_num = (labels != PAD_TOKEN_ID).sum(dim=-1)
-        correct_num = ((labels != PAD_TOKEN_ID) * (final == labels)).sum(dim=-1)
-        cum_acc += torch.true_divide(correct_num, pred_num).mean().item()
-    epoch_loss = running_loss / index
-    epoch_acc = 100 * cum_acc / index
-    epoch_loss = get_reduced_with_fabric(epoch_loss, fabric)
-    epoch_acc = get_reduced_with_fabric(epoch_acc, fabric)
-    #epoch_loss = get_reduced(epoch_loss, local_rank, 0, world_size)
-    #epoch_acc = get_reduced(epoch_acc, local_rank, 0, world_size)
-    epoch_loss
-    if is_master:
-        print(f'    ==  Epoch: {i} | Training Loss: {epoch_loss:.6f} | Accuracy: {epoch_acc:6.4f}%  ==')
-    scheduler.step()
-
-    if i % VALIDATE_EVERY == 0:
-        model.eval()
-        running_loss = 0.0
-        running_error = 0.0
-        predictions = []
-        truths = []
-        with torch.no_grad():
-            for index, data in enumerate(val_loader):
-                print(f'currently in {index}')
-                index += 1
-                data = data.to(device)
-                data, labels = data_mask(data)
-                logits = model(data)
-                loss = loss_fn(logits.transpose(1, 2), labels)
-                running_loss += loss.item()
-                softmax = nn.Softmax(dim=-1)
-                final = softmax(logits)[..., 1:-1]
-                final = final.argmax(dim=-1) + 1
-                predictions.append(final)
-                truths.append(labels)
-            del data, labels, logits, final
-            # gather
-            predictions = distributed_concat(torch.cat(predictions, dim=0), len(val_sampler.dataset), world_size)
-            truths = distributed_concat(torch.cat(truths, dim=0), len(val_sampler.dataset), world_size)
-            correct_num = ((truths != PAD_TOKEN_ID) * (predictions == truths)).sum(dim=-1)[0].item()
-            val_num = (truths != PAD_TOKEN_ID).sum(dim=-1)[0].item()
-            val_loss = running_loss / index
-            val_loss = get_reduced_with_fabric(val_loss, fabric)
-            #val_loss = get_reduced(val_loss, local_rank, 0, world_size)
-        if is_master:
-            val_acc = 100 * correct_num / val_num
-            print(f'    ==  Epoch: {i} | Validation Loss: {val_loss:.6f} | Accuracy: {val_acc:6.4f}%  ==')
-    del predictions, truths
-
-    if is_master:
-        save_ckpt(i, model, optimizer, scheduler, epoch_loss, model_name, ckpt_dir)
-
