--- src/scBERT/requirements.txt
@@ -1,10 +1,0 @@
-einops==0.6.0
-matplotlib<3.7
-numpy==1.20
-pandas==1.1.5
-scanpy==1.7.2
-scikit-learn==0.24.2
-scipy==1.5.4
-torch==1.8.1
-torchvision==0.9.1
-transformers==4.6.1
